Create volume control code
The camera in our device is used for this project. It detects our hand with points in it so as it can see the distance between our thumb finger tip and index finger tip it adjust the device volume accordingly. The distance between the fingers is directly proportional to volume of the device. 

Libraries / Packages used:
cv2
mediapipe
pyautogui